created new file for GoogleMaps feature
seperated googlemap feature from the home page
